% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hub_download.R
\name{Hub_download}
\alias{Hub_download}
\title{API Hub download}
\usage{
Hub_download(xml_file, target_dir, username, password, overwrite = FALSE,
  show_status = FALSE)
}
\arguments{
\item{xml_file}{XML file where query results shall be written to}

\item{username}{API Hub username}

\item{password}{API Hub password}

\item{show_status}{Show (status) messages of the wget download process}
}
\value{
Character vector with (successfully) downloaded file names
}
\description{
Download using wget and Open Search
}
\details{
Download Sentinel-2 (and 1) data from the API Hub
}
\author{
Benjamin Brede
}
\references{
\link{https://scihub.copernicus.eu/twiki/do/view/SciHubUserGuide/5APIsAndBatchScripting#Query_via_Wget}
}

