% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hub_query.R
\name{Hub_query}
\alias{Hub_query}
\title{API Hub query}
\usage{
Hub_query(uri_query, username, password, xml_file = tempfile(fileext =
  ".xml"), show_status = FALSE)
}
\arguments{
\item{uri_query}{URI query as used in the Sentinel SciHub}

\item{username}{API Hub username}

\item{password}{API Hub password}

\item{xml_file}{XML file where query results shall be written to}

\item{show_status}{Show (status) messages of the wget download process}
}
\value{
XML file name (same as input)
}
\description{
Query the API Hub for Sentinel scenes
}
\author{
Benjamin Brede
}
\references{
\link{https://scihub.copernicus.eu/twiki/do/view/SciHubUserGuide/5APIsAndBatchScripting#Query_via_Wget}
}

