% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_bands.R
\name{S2_bands}
\alias{S2_bands}
\title{S2 band names}
\usage{
S2_bands(with_name = NULL, with_resolution = NULL)
}
\arguments{
\item{with_name}{Band names that should be returned. Defaults to all.}

\item{with_resolution}{Resolutions that should be returned. Defaults to all.}
}
\value{
data.frame with available S2 band names and their native resolution
}
\description{
Note that derived bands (all bands not starting with B) are produced at the resolution that is demanded, i.e. don"t have a native resolution
}
\details{
Possible S2 band names and native resolutions
}

