% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_extract_geoinfo.R
\name{S2_extract_geoinfo}
\alias{S2_extract_geoinfo}
\title{Extract geo-location and CRS infos from S2 xml}
\usage{
S2_extract_geoinfo(S2_xml, resolution = c(10, 20, 60))
}
\arguments{
\item{S2_xml}{S2 xml file}

\item{resolution}{For which resolution shall the information be extracted?}
}
\value{
list with dimensions (numeric of 2), UL corner (UTM) cooridnates (numeric of 2), EPSG number of CRS
}
\description{
Extract geo-location and CRS infos from S2 xml
}

