% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_extract_geometry.R
\name{S2_extract_geometry}
\alias{S2_extract_geometry}
\title{Extract sensor geometry}
\usage{
S2_extract_geometry(S2_xml, mode = c("mean", "raster"), ...)
}
\arguments{
\item{S2_xml}{S2 granule xml file}

\item{mode}{Mean values or whole raster for the granule?}

\item{...}{Additional arguments as for \code{\link{writeRaster}}}
}
\value{
Named list (for mode 'mean') or RasterBrickStack (for mode 'raster')
}
\description{
Extract sensor geometry
}
\author{
Benjamin Brede
}

