% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_L1C_granule.R
\name{S2_L1A_granule}
\alias{S2_L1A_granule}
\title{Translate S2 L1C products}
\usage{
S2_L1A_granule(granule_folder, band, lazy = FALSE, ...)
}
\arguments{
\item{granule_folder}{Folder that contains the S2 L2A granule}

\item{band}{Band to extract, can be spectral (B02 to B12 plus B8A) or thematic (SCL or CLD)}

\item{...}{Additional arguments as for \code{\link{writeRaster}}}
}
\value{
RasterLayer
}
\description{
Translate a single S2 L1C granule.
}
\details{
rasterOptions apply for writing the return raster
}

