% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_L2A_granule.R
\name{S2_L2A_granule}
\alias{S2_L2A_granule}
\title{Translate S2 L2A products}
\usage{
S2_L2A_granule(granule_path, band, resolution = c(10, 20, 60), filename,
  overwrite = FALSE, ...)
}
\arguments{
\item{granule_path}{Folder that contains the S2 L2A granule (typically within GRANULE folder and of form S2x_USER_MSI_L2A_..._Nxx.xx)}

\item{band}{Band to extract, can be spectral (B02 to B12 plus B8A) or thematic (SCL or CLD)}

\item{resolution}{Band resolution in m, allowed: 10, 20, 60}

\item{filename}{Output filename; will be automatically suffixed with tileID, band name and resolution}

\item{overwrite}{Overwrite existing files?}

\item{...}{Additional arguments as for \code{\link{gdal_translate}}, not allowed: a_srs, a_ullr.}
}
\value{
RasterLayer
}
\description{
Translate a single S2 L2A granule.
}
\details{
rasterOptions apply for writing the return raster. Filename will be automatically suffixed with granule ID and band name.
}
\author{
Benjamin Brede
}

