% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_L2A_translate.R
\name{S2_L2A_translate}
\alias{S2_L2A_translate}
\title{Translate S2 L2A products}
\usage{
S2_L2A_translate(S2_folder, band, resolution = c(10, 20, 60), filename,
  overwrite = FALSE, ...)
}
\arguments{
\item{S2_folder}{Folder that contains the S2 L2A product (typically suffixed with .SAFE)}

\item{band}{Band to extract, can be spectral (B02 to B12 plus B8A) or thematic (SCL or CLD)}

\item{resolution}{Band resolution in m, allowed: 10, 20, 60}

\item{filename}{Output filename; will be automatically suffixed with tileID, band name and resolution}

\item{overwrite}{Overwrite existing files?}

\item{...}{Additional arguments as for \code{\link{gdal_translate}}, not allowed: a_srs, a_ullr.}
}
\value{
List of RasterLayers corresponding to the raster of the specified band and resolution
}
\description{
Translate an S2 L2A scene into other raster format.
}
\examples{
# select appropriate GDAL installtion that can read JPG2000
gdal_setInstallation('/usr/bin/', rescan = TRUE)

S2 <- 'S2A_USER_PRD_MSIL2A_PDMC_20151001T124224_R008_V20151001T104705_20151001T104705.SAFE'
S2_L2A_translate(S2, 'B8A', 20, 'test.tif')  
}
\author{
Benjamin Brede
}

