% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_gaussian.R
\name{Sensor_gaussian}
\alias{Sensor_gaussian}
\title{Define sensor with gaussian response function}
\usage{
Sensor_gaussian(center, fwhm, band_name, wlunit = "nm", minwl = 400,
  maxwl = 2500, stepsize = 1)
}
\arguments{
\item{center}{Vector with center wavelengths (numeric vector)}

\item{fwhm}{Vector with Full Width Half Maximum values (numeric vector)}

\item{band_name}{Vector with band names (will be coerced to character)}

\item{wlunit}{Label of wavelength unit (character)}

\item{minwl}{Minimum wavelength (numeric)}

\item{maxwl}{Maximum wavelength (numeric)}

\item{stepsize}{Wavelength increment (numeric)}
}
\value{
Data.frame with bands in columns, meta data as attributes
}
\description{
Define sensor with gaussian response function
}
\author{
Benjamin Brede
}

