% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_uniform.R
\name{Sensor_uniform}
\alias{Sensor_uniform}
\title{Define sensor with uniform response function}
\usage{
Sensor_uniform(min, max, band_name, wlunit = "nm", minwl = 400,
  maxwl = 2500, stepsize = 1)
}
\arguments{
\item{min}{Numeric vector with minimum (lower) wavelengths}

\item{max}{Numeric vector with maximum (upper) wavelengths}

\item{band_name}{Vector with band names (will be coerced to character)}

\item{wlunit}{Label of wavelength unit (character)}

\item{minwl}{Minimum wavelength (numeric)}

\item{maxwl}{Maximum wavelength (numeric)}

\item{stepsize}{Wavelength increment (numeric)}
}
\value{
Data.frame with bands in columns, meta data as attributes
}
\description{
Define sensor with uniform response function
}
\author{
Benjamin Brede
}

