% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_L2A_granule.R
\name{S2_L2A_granule}
\alias{S2_L2A_granule}
\title{Translate S2 L2A products}
\usage{
S2_L2A_granule(granule_folder, band, resolution = c(10, 20, 60), ...)
}
\arguments{
\item{granule_folder}{Folder that contains the S2 L2A granule}

\item{band}{Band to extract, can be spectral (B02 to B12 plus B8A) or thematic (SCL or CLD)}

\item{resolution}{Band resolution in m}

\item{...}{Additional arguments as for \code{\link{writeRaster}}}
}
\value{
RasterLayer
}
\description{
Translate a single S2 L2A granule.
}
\details{
rasterOptions apply for writing the return raster
}

