% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_L2A_translate.R
\name{S2_L2A_translate}
\alias{S2_L2A_translate}
\title{Translate S2 L2A products}
\usage{
S2_L2A_translate(S2_folder, band, resolution = c(10, 20, 60), filename, ...)
}
\arguments{
\item{S2_folder}{Folder that contains the S2 L2A product (typically suffixed with .SAFE)}

\item{band}{Band to extract, can be spectral (B01 to B12 plus B8A) or thematic (SCL or CLD)}

\item{resolution}{Band resolution in m}

\item{filename}{Output filename, single tiles will be suffixed with tileID}

\item{...}{Additional arguments as for \code{\link{writeRaster}}}
}
\value{
List of RasterLayers corresponding to the raster of the specified band and resolution
}
\description{
Translate an S2 L2A scene into other raster format.
}
\details{
rasterOptions apply for writing the return rasters
}
\author{
Benjamin Brede
}

