% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_L2A_translate.R
\name{S2_L2A_translate}
\alias{S2_L2A_translate}
\title{Translate S2 L2A products}
\usage{
S2_L2A_translate(S2_folder, band, resolution = c(10, 20, 60), ...)
}
\arguments{
\item{S2_folder}{Folder that contains the S2 L2A product (typically suffixed with .SAFE)}

\item{band}{Band to extract, can be spectral (B01 to B12 plus B8A) or thematic (SCL or CLD)}

\item{resolution}{Band resolution in m}

\item{...}{Additional arguments as for \code{\link{writeRaster}}}
}
\value{
RasterLayer corresponding to the mosaiced raster of the specified band and resolution
}
\description{
Single granules are mosaiced to form one output raster.
}
\details{
Translate a S2 L2A scene into other raster format with CRS.

rasterOptions apply for writing the return raster
}

